// types/PokemonDetail.ts
export interface Ability2 {
  name: string;
  url: string;
}
// ...
export interface Ability3 {
  ability: Ability2;
  is_hidden: boolean;
  slot: number;
}
// ...
export interface Type {
  slot: number;
  type: Type2;
}

export interface Type2 {
  name: string;
  url: string;
}